Set the transforms as per ImageNet requirements
The pre-trained networks were trained on the ImageNet dataset where each
color channel was normalized separately. All three sets, the means, and
standard deviations of the images are normalized to what the network
expects. For the means, it is [0.485, 0.456, 0.406] and for the standard
deviations it is [0.229, 0.224, 0.225], calculated from the ImageNet
images. These values will shift each color channel to be centered at 0
and range from -1 to 1.