Create a pre-trained model with custom classifier
A pre-trained model is created based on any of the torchvision models
available. The model to be used is defined by the value provided in the
'architecture' argument in the object's constructor.

A custom classifier, as per Udacity's project specs, is then set in
place of the selected model's classifier.