Move the model to CPU/GPU as necessary
The model will run in the CPU by default; however, if the model is asked
to run in the GPU, through the 'use_gpu' constructor argument, the model
will be moved to the GPU if one is available. If a GPU is not available
it will still be executed in he CPU and a warning message is displayed.