Set the input tensor size as per selected model
Each torchvision model has a different input tensor size that needs to
be set identical on our custom model's classifier, or the Neural Network
won't work.

Instead of hardcoding the values, the _get_input_units() method returns
the input tensor size based on the original model's classifier. For this
method to work, the model specified needs to be of one of the known base
classes at the moment of this writing:

* AlexNet
* DenseNet
* Inception
* ResNet
* SqueezeNet
* VGG