Set the custom classifier to the correct attribute
Each torchvision model family (AlexNet, DenseNet, VGG, etc.) defines
the classifier in a different variable/attribute name, currently either
defined as 'classifier' or 'fc', which means that the custom classifier
needs to be attached to the correct name. It was previously incorrectly
assumed it was always called 'classifier'.

The custom optimizer also needs to make reference to the proper variable
name or the object constructor might raise an exception.