Add network training validation
During the network training phase, every specified 'x' number of
image batches, the model is validated against the 'valid' image set and
two statistics are collected: 1) the test loss, and 2) the model
accuracy. At present these statistics are informational only and do not
modify the model behavior in any way.

It is expected the test loss to move down, and the accuracy to move up
as the training progresses forward.