Add predict() method
The predict method is passed as argument a path for an image that,
ideally, has not been seen by the model so far and in-turn returns the
following two elements:

1) A list with the 'top_k' highest probabilities, and
2) A list with the class indexes for the highest probabilities above.

Where the 'top_k' number is specified as a method call argument.

The class indexes (item 2 above) are the tensor indexes for the highest
probabilities returned. They don't mean anything by itself and still
need to be mapped to the class names. This mapping will be done at a
later stage.