Read category names JSON file and store in config
The categories are automatically identified by the training algorithm
based on the 'train' sub-directory names from the training dataset.

So, for example, each image under train/1/ directory will be associated
to category 1, each image under train/2/ directory under category 2,
each image under train/125/ directory to category 125, and so on.

These category 'names' might not be descriptive enough. The purpose of
the 'category names' JSON file is to provide a mapping between the
category IDs identified by the training algorithm (based on directory
names), and a descriptive category name that can be then used for
reporting purposes.