Store dataset.class_to_idx in object config
After the classifier has been trained, it is necessary to store in the
object's config attribute the 'train' dataset's class_to_idx contents.
This allows for later mapping of class IDs, as retrieved from image
directories, to the classes' descriptive names as provided by the
category names JSON file, if one was provided at object construction.